Refactor codeforces client to use the api instead of web scraping
Previously web scraping was used for fetching submission data, most
importantly the submission code, this no longer works as the HTML pages
fetched does not contain the source code for submissions, which may be
due to anti-scraping measures by cf or js rendering (not sure).

Either way cf explicitly disallows every form of scraping, as stated in
their `robots.txt`:

  User-agent: python-requests
  Disallow: /
  Allow: /api/

And it would be harder to construct a solution around to retrieve the
code with scraping.

I propose a migration to the cf api in order to avoid the mentioned
problems and comply with their scraping policy. The api has very few
endpoints and some limited information (I dit not found contest tags),
but the core information needed (source code) is available.

Platform configurations are now dictiionaries with their corresponding
properties, which have at least "username".